refactor(botonic-react): create webchat context folder
