+++ packages/botonic-react/src/components/carousel.jsx
@@ -5,8 +5,8 @@ import styled from 'styled-components'
 import LeftArrow from '../assets/leftArrow.svg'
 import RightArrow from '../assets/rightArrow.svg'
 import { COLORS, WEBCHAT } from '../constants'
-import { WebchatContext } from '../contexts'
 import { resolveImage } from '../util/environment'
+import { WebchatContext } from '../webchat/context'
 import { ButtonsDisabler } from './buttons-disabler'
 import { Message } from './message'
