+++ packages/botonic-react/src/contexts.tsx
@@ -1,8 +1,7 @@
 import { Input as CoreInput, Session as CoreSession } from '@botonic/core'
 import { createContext } from 'react'
 
-import { ActionRequest, WebchatContextProps } from './index-types'
-import { webchatInitialState } from './webchat/hooks'
+import { ActionRequest } from './index-types'
 
 export const RequestContext = createContext<
   Partial<ActionRequest> & {
@@ -35,78 +34,4 @@ export const WebviewRequestContext = createContext<{
   getString: undefined,
   params: {} as Record<string, any>,
   session: {} as Partial<CoreSession>,
-})
-
-export const WebchatContext = createContext<WebchatContextProps>({
-  addMessage: () => {
-    return
-  },
-  getThemeProperty: () => {
-    return
-  }, // used to retrieve a specific property of the theme defined by the developer in his 'webchat/index.js'
-  openWebview: () => {
-    return
-  },
-  resolveCase: () => {
-    return
-  },
-  resetUnreadMessages: () => {
-    return
-  },
-  setIsInputFocused: () => {
-    return
-  },
-  setLastMessageVisible: () => {
-    return
-  },
-  sendAttachment: async () => {
-    return
-  },
-  sendInput: async () => {
-    return
-  },
-  sendPayload: async () => {
-    return
-  },
-  sendText: async () => {
-    return
-  },
-  theme: {},
-  toggleCoverComponent: () => {
-    return
-  },
-  toggleWebchat: () => {
-    return
-  },
-  toggleEmojiPicker: () => {
-    return
-  },
-  togglePersistentMenu: () => {
-    return
-  },
-  updateLatestInput: () => {
-    return
-  },
-  updateMessage: () => {
-    return
-  },
-  updateReplies: () => {
-    return
-  },
-  updateUser: () => {
-    return
-  },
-  updateWebchatDevSettings: () => {
-    return
-  },
-  webchatState: webchatInitialState,
-  trackEvent: async () => {
-    return
-  },
-  webchatRef: { current: null },
-  chatAreaRef: { current: null },
-  inputPanelRef: { current: null },
-  headerRef: { current: null },
-  scrollableMessagesListRef: { current: null },
-  repliesRef: { current: null },
 })
