+++ packages/botonic-react/src/webchat/header/default-header.tsx
@@ -1,10 +1,10 @@
 import React, { useContext } from 'react'
 
 import { COLORS, ROLES, WEBCHAT } from '../../constants'
-import { WebchatContext } from '../../contexts'
 import { Scale } from '../../shared/styles'
 import { resolveImage } from '../../util/environment'
 import { ConditionalWrapper } from '../../util/react'
+import { WebchatContext } from '../../webchat/context'
 import {
   CloseHeader,
   HeaderContainer,
