+++ packages/botonic-react/src/webchat/hooks/index.ts
@@ -6,6 +6,5 @@ export { useScrollToBottom } from './use-scroll-to-bottom'
 export { useScrollbarController } from './use-scrollbar-controller'
 export { useTyping } from './use-typing'
 export { useVirtualKeyboardDetection } from './use-virtual-keyboard-detection'
-export { useWebchat, webchatInitialState } from './use-webchat'
 export { useWebchatDimensions } from './use-webchat-dimensions'
 export { useWebchatResizer } from './use-webchat-resizer'
