+++ packages/botonic-react/src/webchat/hooks/use-scrollbar-controller.ts
@@ -1,6 +1,6 @@
 import { useContext, useEffect } from 'react'
 
-import { WebchatContext } from '../../contexts'
+import { WebchatContext } from '../../webchat/context'
 import { DEVICES, isMobileDevice } from '../devices'
 
 // TODO: Investigate why when we have some messages, scroll actions are not disabled properly
