+++ packages/botonic-react/src/webchat/index.ts
@@ -1,4 +1,5 @@
 // @ts-nocheck
+export { WebchatContext } from './context'
 export { Webchat } from './webchat'
 export { WebchatDev } from './webchat-dev'
