+++ packages/botonic-react/src/webchat/input-panel/persistent-menu.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react'
 
 import LogoMenu from '../../assets/menuButton.svg'
 import { ROLES, WEBCHAT } from '../../constants'
-import { WebchatContext } from '../../contexts'
+import { WebchatContext } from '../../webchat/context'
 import { Icon } from '../components/common'
 import { ConditionalAnimation } from '../components/conditional-animation'
