+++ packages/botonic-react/tests/helpers/test-utils.jsx
@@ -4,7 +4,7 @@ import TestRenderer from 'react-test-renderer'
 
 import { MultichannelContext } from '../../src/components/multichannel/multichannel-context'
 import { RequestContext } from '../../src/contexts'
-import { useWebchat } from '../../src/webchat/hooks'
+import { useWebchat } from '../../src/webchat/context/use-webchat'
 
 /**
  *
